refactor: seperate cart reducer function
