+++ src/reducers/index.js
@@ -1,0 +1,2 @@
+export {reducer} from "./product-reducer"
+export {cartReducer} from "./cart-reducer"
