+++ src/utils/functions/calculatePercentage.js
@@ -1,6 +1,3 @@
-
-
-
 export const calculatePercentage = (price) => {
     const percentage =
         ((price.oldPrice - price.newPrice) * 100) / price.oldPrice;
